Triton submarines carry too much ammo!
Because the Mk51 is a new design, it can often be difficult for Triton
captains to get their full complement of 44 weapons. Often, submarines
put to see carrying a large number of ADCAP or CBASS torpedoes in
addition to their Mk51s. The big problem there is that these torpedoes
are usually incompatible with Triton's swim-free tubes and, therefore,
will produce a launch transient.